Create a file for Loader class
